fix: `getUserById`もロジック分割してそれ以外は`use server`削除
