+++ data/user.ts
@@ -1,4 +1,3 @@
-"use server"
 import { db } from "@/utils/db"
 
 export const getUserByEmail = async (email: string) => {
@@ -11,31 +10,6 @@ export const getUserByEmail = async (email: string) => {
     return user
   } catch (error) {
     console.error("getUserByEmail: ", error)
-    return null
-  }
-}
-
-export const getUserById = async (id: string) => {
-  try {
-    const user = await db.user.findUnique({
-      where: {
-        id,
-      },
-      select: {
-        id: true,
-        name: true,
-        email: true,
-        profileText: true,
-        studentNumber: true,
-        updatedAt: true,
-        createdAt: true,
-        iconImagePath: true,
-        accounts: true,
-      },
-    })
-    return user
-  } catch (error) {
-    console.error("getUserById: ", error)
     return null
   }
 }
